third_party/blink/web_tests/external/wpt/css/css-transitions/transitionevent-interface-expected.txt

This is a testharness.js-based test.
[FAIL] TransitionEventInit animation property sets value
  assert_equals: expected (object) object "[object CSSTransition]" but got (undefined) undefined
Harness: the test ran to completion.
